Exclude certain findbugs warnings
Exclude warnings about exposing objects in arrays. Since we are not
creating an API, but rather a game as a whole, we can trust that the
arrays are not modified by the providing class (unless if there is a good
reason to do so). Copying the arrays would cause very large additional
memory usage without adding any functionality or security.

Exclude warnings about unusual return results like File.delete(). We want
to explicitly ignore failed deletions, as failing file deletion should not
interrupt the gameplay in any way. If the file deletion failure was
important in any way, then the error will occur when that file is
attempted to be overridden elsewhere, which is a more applicable place for
the error.